initial_material_ui added drawer and grid
The basics are in place for our application. The drawer will serve our
content as if it is a mobile app with dynamic content showing up in the
body immediately to the right that responds to "get out of the way"
geometric changes. The remainder of the grid will be addressed at a
later time. There is boiler plate routing for a SPA in place, but
everything is commented out because there is no need to do any of that
right now.
