+++ src/gridNexus.js
@@ -1,0 +1,88 @@
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import Paper from '@material-ui/core/Paper';
+import Grid from '@material-ui/core/Grid';
+import Gdrawer from './gdrawer';
+import Home from './Home';
+import Stuff from './Stuff';
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    flexGrow: 1,
+  },
+  paper: {
+    padding: theme.spacing(2),
+    textAlign: 'center',
+    color: theme.palette.text.secondary,
+  },
+}));
+
+export default function CenteredGrid() {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.root}>
+      <Grid
+        container
+        spacing={3}
+      >
+        <Grid
+          item
+          xs={12}
+        >
+          <Paper
+            className={classes.paper}
+          > <Gdrawer /> </Paper>
+        </Grid>
+        <Grid
+          item
+          xs={6}
+        >
+          <Paper
+            className={classes.paper}
+          > <Home /> </Paper>
+        </Grid>
+        <Grid
+          item
+          xs={6}
+        >
+          <Paper
+            className={classes.paper}
+          > <Stuff /> </Paper>
+        </Grid>
+        <Grid
+          item
+          xs={3}
+        >
+          <Paper
+            className={classes.paper}
+          >xs=3</Paper>
+        </Grid>
+        <Grid
+          item
+          xs={3}
+        >
+          <Paper
+            className={classes.paper}
+          >xs=3</Paper>
+        </Grid>
+        <Grid
+          item
+          xs={3}
+        >
+          <Paper
+            className={classes.paper}
+          >xs=3</Paper>
+        </Grid>
+        <Grid
+          item
+          xs={3}
+        >
+          <Paper
+            className={classes.paper}
+          >xs=3</Paper>
+        </Grid>
+      </Grid>
+    </div>
+  );
+}
